+++ database/migrations/2016_03_21_085609_create_hyperlinks_table.php
@@ -16,7 +16,7 @@ class CreateHyperlinksTable extends Migration
             $table->bigIncrements('id');
             $table->string('name')->nullable();
             $table->string('link', 2083);
-            $table->integer('project_id')->index()->unsigned();
+            $table->bigInteger('project_id')->index()->unsigned();
             $table->dateTime('created_at');
             $table->dateTime('updated_at');
